added all components required to assign Trainee to Training
Notice how the responsibility of saving new TrainingAssignment is thrown out of the Training class. Simply adding new assignemnt to "_trainees" collection would be "nicer", but would also cause NH fetch all already assigned trainees - we do not need that here.
Instead we mark this realtion as "inverse" in NH and let command handler save the new object.
Of course lots of validation and user-friendly error messages are missing at this point, but it does not matter.